V4/s2p_v4_server/requirements.txt
opencv-contrib-python==4.1.0.25
tensorflow_gpu==1.14.0
bottle==0.12.10
keras==2.2.5
scikit-learn==0.23.1
scikit-image==0.14.5
llvmlite==0.36.0rc1
numba==0.53.1
tqdm
paste

